+++ abides-jpmc-public/abides_markets/utils/__init__.py
@@ -162,7 +162,7 @@ def generate_latency_model(agent_count, latency_type="deterministic"):
         "no_latency",
     ], "Please select a correct latency_type"
 
-    latency_rstate = np.random.RandomState(seed=np.random.randint(low=0, high=2 ** 32))
+    latency_rstate = np.random.RandomState(seed=np.random.randint(low=0, high=2 ** 32, dtype="uint64"))
     pairwise = (agent_count, agent_count)
 
     if latency_type == "deterministic":
